feat: changed previous boarding layout to screens layout and added SafeArea widget (default layout on all screens) & removed unnecessary scaffold widgets making the app lighter
