+++ client/lib/pages/login/help_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 // Widgets & Layout
-import 'package:flutter_application/layouts/boarding_layout.dart';
+import 'package:flutter_application/layouts/screens_layout.dart';
 import 'package:flutter_application/utils/helpers/app_layout.dart';
 import 'package:flutter_application/utils/styles/typography.dart';
 import 'package:flutter_application/widgets/app_layout/app_screen_title.dart';
@@ -16,48 +16,46 @@ class HelpPage extends StatelessWidget {
   Widget build(BuildContext context) {
     final dynamic size = AppLayout.getSize(context);
 
-    return Scaffold(
-      body: BoardingLayout(
-        column: Column(
-          children: [
-            AppScreenTitle(
-              text: "How to login",
-              leftButton: IconButton(
-                onPressed: () => Navigator.pop(context),
-                icon: const Icon(
-                  Icons.arrow_back_ios,
-                  size: 20,
-                  color: Colors.white,
+    return ScreensLayout(
+      column: Column(
+        children: [
+          AppScreenTitle(
+            text: "How to login",
+            leftButton: IconButton(
+              onPressed: () => Navigator.pop(context),
+              icon: const Icon(
+                Icons.arrow_back_ios,
+                size: 20,
+                color: Colors.white,
+              ),
+            ),
+          ),
+          Container(
+            padding: const EdgeInsets.symmetric(horizontal: 25, vertical: 25),
+            margin: const EdgeInsets.only(top: 30),
+            decoration: BoxDecoration(
+              color: AppColors.whiteColor,
+              borderRadius: const BorderRadius.all(Radius.circular(20)),
+            ),
+            width: size.width * 0.9,
+            height: size.height * 0.8,
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  "TBC",
+                  style: AppTypography.largeTitle
+                      .copyWith(color: AppColors.smokyBlackColor),
                 ),
-              ),
+                Text(
+                  "Content",
+                  style: AppTypography.textBody
+                      .copyWith(color: AppColors.smokyBlackColor),
+                ),
+              ],
             ),
-            Container(
-              padding: const EdgeInsets.symmetric(horizontal: 25, vertical: 25),
-              margin: const EdgeInsets.only(top: 30),
-              decoration: BoxDecoration(
-                color: AppColors.whiteColor,
-                borderRadius: const BorderRadius.all(Radius.circular(20)),
-              ),
-              width: size.width * 0.9,
-              height: size.height * 0.8,
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text(
-                    "TBC",
-                    style: AppTypography.largeTitle
-                        .copyWith(color: AppColors.smokyBlackColor),
-                  ),
-                  Text(
-                    "Content",
-                    style: AppTypography.textBody
-                        .copyWith(color: AppColors.smokyBlackColor),
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
+          ),
+        ],
       ),
     );
   }
